requirements.txt
alpaca-trade-api
streamlit
pandas
plotly
python-dotenv
schedule
polygon-api-client
websockets
